requirements.txt
django>=2.0.8
channels>=2.0.0
redis
channels_redis
djangorestframework
pyfcm
asgiref

psycopg2-binary
